moved bottom buttons from home to defaultlayout
